Fix generation of kernel change uevents for partitions after volume rename on linux
The current method of tickling the disk readonly attribute will only
generate a change uevent for the /dev/zdNNNN device, any partitions
which have been presented as /dev/zdNNNNpX do not have an event
generated.  This will leave incorrectly named symlinks at
/dev/zvol/ztank/zvol-partX.

Iterate over all the partitions of the disk and trigger a change event
for all of the devices so that udev updates the symlinks as expected.

Signed-off-by: James Dingwall <james@dingwall.me.uk>